x86_64 __redirect_ieee754_expf: Change double to float
__redirect_ieee754_expf has type float, not double.

	* sysdeps/x86_64/fpu/multiarch/e_expf.c (__redirect_ieee754_expf):
	Change double to float.

+++ sysdeps/x86_64/fpu/multiarch/e_expf.c
@@ -16,7 +16,7 @@
    License along with the GNU C Library; if not, see
    <http://www.gnu.org/licenses/>.  */
 
-extern double __redirect_ieee754_expf (double);
+extern float __redirect_ieee754_expf (float);
 
 #define SYMBOL_NAME ieee754_expf
 #include "ifunc-fma.h"
